+++ core/src/codegen/variant_data.rs
@@ -40,8 +40,17 @@ impl<'a> FieldsGen<'a> {
         let handle_unknown = if self.allow_unknown_fields {
             quote!()
         } else {
+            // We can't call `unknown_field_with_alts` with an empty slice, or else it fails to
+            // infer the type of the slice item.
+            let err_fn = if arms.is_empty() {
+                quote!(unknown_field(__other))
+            } else {
+                let names = self.fields.as_ref().map(Field::as_name);
+                quote!(unknown_field_with_alts(__other, &[#(#names),*]))
+            };
+
             quote! {
-                __errors.push(::darling::Error::unknown_field(__other).with_span(__inner));
+                __errors.push(::darling::Error::#err_fn.with_span(__inner));
             }
         };